✨🐛 Add DevCyclePR workflow config
This commit adds the DevCyclePR workflow configuration file to the
`.github/workflows` directory. The workflow is triggered by a
`workflow_call` event and consists of a single job named
`dvc-feature-flag-insights`. This job runs on `ubuntu-latest` and is
responsible for generating DevCycle PR insights. The job includes
steps to checkout the code and utilize the
`DevCycleHQ/feature-flag-pr-insights-action@v2.0.0` action. The action
requires several secrets to be provided in order to function
properly.